Move fastapprox to include/meta/math.
