+++ include/meta/math/fastapprox.h
