+++ src/index/ranker/okapi_bm25.cpp
@@ -7,7 +7,7 @@
 #include "meta/index/inverted_index.h"
 #include "meta/index/ranker/okapi_bm25.h"
 #include "meta/index/score_data.h"
-#include "meta/util/fastapprox.h"
+#include "meta/math/fastapprox.h"
 
 namespace meta
 {
